:hammer: Refactor business logic for token module methods
