+++ services/blockchain-indexer/shared/dataService/business/token/index.js
@@ -1,0 +1,30 @@
+/*
+ * LiskHQ/lisk-service
+ * Copyright © 2023 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ *
+ */
+const { tokenHasUserAccount } = require('./accountExists');
+const { getAvailableTokenIDs } = require('./availableIDs');
+const { getTokenBalances } = require('./balances');
+const { getTokenConstants } = require('./constants');
+const { getTokenSummary } = require('./summary');
+const { getTokenTopBalances } = require('./topBalances');
+
+module.exports = {
+	tokenHasUserAccount,
+	getAvailableTokenIDs,
+	getTokenBalances,
+	getTokenConstants,
+	getTokenSummary,
+	getTokenTopBalances,
+};
