+++ services/blockchain-indexer/shared/dataService/business/token/summary.js
@@ -1,0 +1,65 @@
+/*
+ * LiskHQ/lisk-service
+ * Copyright © 2022 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ *
+ */
+const { requestConnector } = require('../../../utils/request');
+
+const {
+	LENGTH_CHAIN_ID,
+	PATTERN_ANY_TOKEN_ID,
+	PATTERN_ANY_CHAIN_TOKEN_ID,
+} = require('../../../constants');
+
+const getTokenSummary = async () => {
+	const summary = {
+		data: {},
+		meta: {},
+	};
+
+	const { escrowedAmounts } = await requestConnector('getEscrowedAmounts');
+	const { totalSupply } = await requestConnector('getTotalSupply');
+	const { supportedTokens: supportedTokenIDs } = await requestConnector('getSupportedTokens');
+
+	const supportedTokens = {
+		isSupportAllTokens: false,
+		exactTokenIDs: [],
+		patternTokenIDs: [],
+	};
+
+	supportedTokenIDs.forEach(tokenID => {
+		if (tokenID === PATTERN_ANY_TOKEN_ID) {
+			supportedTokens.isSupportAllToken = true;
+		} else if (tokenID.substring(LENGTH_CHAIN_ID) === PATTERN_ANY_CHAIN_TOKEN_ID) {
+			supportedTokens.patternTokenIDs.push(tokenID);
+		} else {
+			supportedTokens.exactTokenIDs.push(tokenID);
+		}
+	});
+
+	summary.data = {
+		escrowedAmounts,
+		supportedTokens: {
+			...supportedTokens,
+			exactTokenIDs: [...new Set(supportedTokens.exactTokenIDs)],
+			patternTokenIDs: [...new Set(supportedTokens.patternTokenIDs)],
+		},
+		totalSupply,
+	};
+
+	return summary;
+};
+
+module.exports = {
+	getTokenSummary,
+};
